сделан модульный js
